# app/backend/docs/swagger.yaml
basePath: /
definitions:
...
  domain.CategoryUpsert:
    properties:
      id:
        type: string
      name_i18n:
        type: object
      parent_id:
        type: string
      slug:
        type: string
    type: object
...
  domain.EvidenceUpsert:
    properties:
      id:
        type: string
      media_type:
        type: string
      object_key:
        type: string
      observation_id:
        type: string
      observation_type:
        type: string
      sha256:
        type: string
      uploaded_by:
        type: string
    type: object
  domain.FuelPriceUpsert:
    properties:
      amount_per_litre:
        type: string
      confidence_score:
        type: string
      contributor_id:
        type: string
      currency:
        type: string
      fuel_type_id:
        type: string
      id:
        type: string
      location_id:
        type: string
      observed_at:
        type: string
      source_id:
        type: string
      source_record_id:
        type: string
      status:
        type: string
    type: object
  domain.FuelTypeUpsert:
    properties:
      code:
        type: string
      energy:
        type: string
      id:
        type: string
      name_i18n:
        type: object
    type: object
  domain.GeoAreaUpsert:
    properties:
      code:
        type: string
      country_code:
        type: string
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      parent_id:
        type: string
      type:
        type: string
    type: object
...
  httpapi.Problem:
    properties:
      detail:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  httpapi.ResourceResponse:
    properties:
      data: {}
    type: object
...
info:
  contact: {}
  description: |-
    API de comparaison du coût de la vie dans la zone euro.
    Les montants décimaux sont exposés sous forme de chaînes pour préserver leur précision.
  title: OjoZone API
  version: "1.0"
paths:
  /api/v1/admin/users:
    get:
      description: Lecture seule en l'absence d'authentification ; password_hash n'est
        jamais exposé.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Consulter les utilisateurs en administration
      tags:
      - Administration
  /api/v1/admin/users/{id}:
    get:
      description: Lecture seule en l'absence d'authentification ; password_hash n'est
        jamais exposé.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Consulter les utilisateurs en administration
      tags:
      - Administration
...
  /api/v1/categories:
    get:
      consumes:
      - application/json
      parameters:
      - description: Catégorie
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.CategoryUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les catégories
      tags:
      - Catégories
    post:
      consumes:
      - application/json
      parameters:
      - description: Catégorie
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.CategoryUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les catégories
      tags:
      - Catégories
  /api/v1/categories/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Catégorie
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.CategoryUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les catégories
      tags:
      - Catégories
    get:
      consumes:
      - application/json
      parameters:
      - description: Catégorie
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.CategoryUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les catégories
      tags:
      - Catégories
    put:
      consumes:
      - application/json
      parameters:
      - description: Catégorie
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.CategoryUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les catégories
      tags:
      - Catégories
  /api/v1/evidence-files:
    get:
      consumes:
      - application/json
      parameters:
      - description: Preuve
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.EvidenceUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les preuves
      tags:
      - Preuves
    post:
      consumes:
      - application/json
      parameters:
      - description: Preuve
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.EvidenceUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les preuves
      tags:
      - Preuves
  /api/v1/evidence-files/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Preuve
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.EvidenceUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les preuves
      tags:
      - Preuves
    get:
      consumes:
      - application/json
      parameters:
      - description: Preuve
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.EvidenceUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les preuves
      tags:
      - Preuves
    put:
      consumes:
      - application/json
      parameters:
      - description: Preuve
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.EvidenceUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les preuves
      tags:
      - Preuves
  /api/v1/fuel-prices:
    get:
      consumes:
      - application/json
      parameters:
      - description: Observation
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.FuelPriceUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les prix des carburants
      tags:
      - Prix carburants
    post:
      consumes:
      - application/json
      parameters:
      - description: Observation
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.FuelPriceUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les prix des carburants
      tags:
      - Prix carburants
  /api/v1/fuel-prices/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Observation
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.FuelPriceUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les prix des carburants
      tags:
      - Prix carburants
    get:
      consumes:
      - application/json
      parameters:
      - description: Observation
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.FuelPriceUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les prix des carburants
      tags:
      - Prix carburants
    put:
      consumes:
      - application/json
      parameters:
      - description: Observation
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.FuelPriceUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les prix des carburants
      tags:
      - Prix carburants
  /api/v1/fuel-types:
    get:
      consumes:
      - application/json
      parameters:
      - description: Type de carburant
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.FuelTypeUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les types de carburant
      tags:
      - Carburants
    post:
      consumes:
      - application/json
      parameters:
      - description: Type de carburant
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.FuelTypeUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les types de carburant
      tags:
      - Carburants
  /api/v1/fuel-types/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Type de carburant
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.FuelTypeUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les types de carburant
      tags:
      - Carburants
    get:
      consumes:
      - application/json
      parameters:
      - description: Type de carburant
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.FuelTypeUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les types de carburant
      tags:
      - Carburants
    put:
      consumes:
      - application/json
      parameters:
      - description: Type de carburant
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.FuelTypeUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les types de carburant
      tags:
      - Carburants
  /api/v1/geo-areas:
    get:
      consumes:
      - application/json
      parameters:
      - description: Zone géographique
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.GeoAreaUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les zones géographiques
      tags:
      - Zones géographiques
    post:
      consumes:
      - application/json
      parameters:
      - description: Zone géographique
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.GeoAreaUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les zones géographiques
      tags:
      - Zones géographiques
...
  /api/v1/housing-observations:
    get:
      consumes:
      - application/json
      parameters:
      - description: Observation
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.HousingUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les observations immobilières
      tags:
      - Immobilier
    post:
      consumes:
      - application/json
      parameters:
      - description: Observation
        in: body
        name: payload
        schema:
          $ref: '#/definitions/domain.HousingUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.Problem'
      summary: Gérer les observations immobilières
      tags:
      - Immobilier
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"